+++ src/utils/fetch/never-throwing-fetch.test.ts
@@ -3,14 +3,17 @@ import { neverThrowingFetch } from "./never-throwing-fetch"
 
 describe("neverThrowingFetch", () => {
   it("should call fetch function", async () => {
-    vi.spyOn(window, "fetch")
+    vi.spyOn(globalThis, "fetch")
     const options: RequestInit = { credentials: "include" }
     await neverThrowingFetch("https://example.com", options)
-    expect(window.fetch).toHaveBeenCalledWith("https://example.com", options)
+    expect(globalThis.fetch).toHaveBeenCalledWith(
+      "https://example.com",
+      options,
+    )
   })
 
   it("should return response", async () => {
-    vi.spyOn(window, "fetch").mockImplementation(() =>
+    vi.spyOn(globalThis, "fetch").mockImplementation(() =>
       Promise.resolve(new Response()),
     )
     const response = await neverThrowingFetch("https://example.com")
@@ -18,7 +21,7 @@ describe("neverThrowingFetch", () => {
   })
 
   it("should return an error on fetch failure", async () => {
-    vi.spyOn(window, "fetch").mockImplementation(() =>
+    vi.spyOn(globalThis, "fetch").mockImplementation(() =>
       Promise.reject(new Error("Fetch failed")),
     )
     const response = await neverThrowingFetch("https://example.com")
@@ -27,7 +30,9 @@ describe("neverThrowingFetch", () => {
 
   it("should return an error with the original error as cause", async () => {
     const error = new Error("Fetch failed")
-    vi.spyOn(window, "fetch").mockImplementation(() => Promise.reject(error))
+    vi.spyOn(globalThis, "fetch").mockImplementation(() =>
+      Promise.reject(error),
+    )
     const response = await neverThrowingFetch("https://example.com")
     expect((response as Error).cause).toBe(error)
   })